Move proofreader.
