+++ packages/core/src/components/Proofreader/types.ts
