+++ packages/core/src/components/TextArea/index.tsx
@@ -5,8 +5,8 @@ import { requiredBy } from 'airbnb-prop-types';
 import BaseTextArea, { Props as BaseTextAreaProps } from '../private/BaseTextArea';
 import FormField, { Props as FormFieldProps, partitionFieldProps } from '../FormField';
 import T from '../Translate';
-import Proofreader, { Props as ProofreaderProps } from './Proofreader';
-import { ExtraProofreadProps } from './Proofreader/types';
+import Proofreader, { Props as ProofreaderProps } from '../Proofreader';
+import { ExtraProofreadProps } from '../Proofreader/types';
 
 export type Props = Omit<BaseTextAreaProps, 'id'> &
   FormFieldProps &
